requirements.txt
requests
python-dotenv
web3==5.31.4
